Favorability: consolidate 10 modules into 7 (pure code motion)
Fold the two extraction-only satellites into extract.py and the argparse
entry into the runner, mirroring the main scraper's flat layout:

- extract.py absorbs derive.py (Ipsos index arithmetic, expressers-rate
  derivation) and validate.py (LLM payload schema check, row guards); both
  existed only to serve extraction. Imports repointed; delete derive.py,
  validate.py.
- crawler.py absorbs cli.py (argparse main + __main__), like daily_update.py
  is both runner and entrypoint. Delete cli.py; update the daily GitHub
  Action step and readme to the crawler module path.

No logic touched: classification tree, taxonomy, base invariant, derive
arithmetic and calendar-decay averaging are unchanged. favorability_polls
.csv/.jsonl and favorability_averages.csv stay byte-identical; 174 tests
green; offline reprocess with OPENAI_API_KEY unset unaffected.
